Library (fits.h): accounts for 3 number version string for CFITSIO
From CFITSIO version '4.0.0', CFITSIO has changed its version
format. Before this, it was a 2 digit number (like '3.49' or '3.48' and
etc), but from this major version, it will have three digits!

Because of the old format, until now, Gnuastro was printing the version as
a floating point number (with two decimal point precision). As a result,
the new versioning format of CFITSIO (which is no longer a floating point
number, and has three '.'s) caused a crash when building Gnuastro. But
fortunately, with the new number in the version, CFITSIO has also
introduced the 'CFITSIO_MICRO' macro to keep it as an integer.

With this commit, a pre-processor check has been added to see if
'CFITSIO_MICRO' is defined or not. If its not defined, then the user has an
older version of CFITSIO and Gnuastro will print the version as a 2-integer
number (separated by a '.'). If it is defined, then a three-digit version
number is printed.

This bug was reported by Vincenzo Testa and Zohreh Ghaffari.

This fixes bug #61007.