NEWS: acknowledged the source of the bug fixed in previous commit
In the previous commit, I forgot to acknowledge Vincenzo Testa and Zohreh
Ghaffari who reported the bug that was fixed there in the NEWS file. Since
this file is included in public releases, it is important to properly
acknowledge all those who contributed.

With this commit, they are mentioned as the reporters of that bug.