Crop: using library to write crop information FITS keywwords
Until now, Crop done the low-level writing of the the FITS keyword title
for "Crop information" keywords itself! As a result, it didn't insert an
empty line before the new title and it would be immediately in the line
after the WCS! This makes it hard to notice and didn't follow the regular
pattern of one empty lines between each titled group of keywords.

With this commit, Crop now uses the proper Gnuastro library function for
writing FITS keyword titles (that is also used by other programs). This
fixes the problem above, and also simplified the coding (many extra lines
have been removed).