PSF select stars: remove --input add --output and --catalog
Until now, some parameters such as (the old) '--matchaperture' and
'--minaxisrario' mistakenly had been written. Furthermore for inserting
image we had a '--input' option, while in Gnuastro we do not have any
option like that. We did not use '--output' for defining output of script,
also '--extscript' option was extra.

Also, this script is just used when the user has an image. Based on that
image and by using 'astquery', this script downloads the corresponding
catalog of stars from the Gaia data base. Then based on this catalog
circular stars with good parallax and less than 9 neighbors were
selected. If the user has already prepared the catalog and does not want to
download it again, or the catalog of stars is not in Gaia, for instance in
simulated images, this pipeline could not be used.

With this commit, the parameters that have been written mistakenly are
corrected. For instance, instead of '--matchaperture' we now have
'--matchaperturedeg' and instead of '--minaxisrario' we now have
'--minaxisratio'. The '--input' option also has been removed and we now
have a new '--output' option for the result of the script, also
'--extscript' was removed.

Furthermore, if the user has prepared the catalog before or has a simulated
images that won't be in Gaia s/he can use the new '--catalog' option. With
this option, at first this script checks the overlap with the image and
aborts with an error if there is not any overlap.