Library (txt.h): other metadata used when type can't be read
Until now, when the input plain-text table's metadata contains a type
identifier that is not understood, Gnuastro's plain-text table library
would ignore the full metadata. However, the numeric data type is an
optional metadata, and it if doesn't exist, Gnuastro will use a default
data type (float64). So if it is not understood (usually due to a typo), it
is unexpected to ignore the metadata over all, we should just use the
default type.

With this commit, this change has been implemented and the description of
Gnuastro's plain-text table format has been updated correspondingly.

This bug was reported by Sepideh Eskandarlou.

This fixes bug #61108.