Book: minor edits to the description of gal_arithmetic
In the previous commit, the description of this function was edited to be
more clear. However, Raul Infante-Sainz reported that instead of a base-e,
I had written base-2 for the 'log' function! After looking through it
again, some other edits occurred to me that could futher help clarify the
point.

With this commit, it has been edited again, and in the process, the
description of the 'GAL_ARITHMETIC_OP_LOG' operator (as well as the 'sqrt'
and 'log10' operators) were also edited to be more clear.