Book: better description --width giving odd-sided crop
Until now, the book briefly discussed this feature of the '--width' option
(in combination with '--center') and even proposed a solution (on how to
make it even-sided if necessary). But that description wasn't enough and
could still lead to confusions.

With this commit, to better clarify the situation, a paragraph has been
added with a more precise description of the reason behind this behavior.

This was suggested by Alejandro Serrano Borlaff.