PSF select stars: if the image has not stars in `--min` and `--max` range of magnitude
Until now, if the user had been specified to select the stars in the range
of magnitude between the `--min` to `--max`, while in the image, stars in
the range of magnitude between `--min` and `--max` was not exist, the error
of the "--stdintimeout" had been ocured. Without any warning about the
stars in this range of magnitude was not exist.

Now with this commit, when the user has specified to select the stars in
the range of magnitude between `--min`-`--max`, while in the image any
stars in this range of magnitude does not exist, the script will be
stopped, and print the "There were no stars in magnitude range `--min` to
`--max`". In this case the user will be know where is the problem.