Build system: more clear message when man pages come from tarball
When the Help2Man program isn't available on the host, Gnuastro will assume
that they are already present in the tarball and will attempt to copy
them. However, the message (that informs the user about this) was just
saying "Using distributed man page for ...", which was very ambiguous and
caused us some confusion in a recent debugging issue: 'help2man' wasn't
installed when Gnuastro was configured, as a result, it would print this
message and not build the man pages.

With this commit, to be more clear the message now reads "Using man page in
tarball (no help2man) for ...".

This issue was found and fixed with the help of Raul Infante-Sainz.