BuildProgram: all necessary linking flags now added
Until now, when constructing the command to build custom programs,
BuildProgram wouldn't include the commands to link with Gnuastro's
dependencies. This was because all those dependencies were already written
into Gnuastro's libtool file (the installed 'libgnuastro.la' file that
BuildProgram uses). However, as part of bug #61192, we discovered that
Debian (and thus all derivate OSs) apply patches to Libtool that disable
this very useful behavior (to avoid complicating their packaging). As a
result, building of custom programs with BuildProgram would fail due to
linking errors when external libraries were necessary (for example 'sin' in
the C math library: '-lm').

With this commit, BuildProgram now includes all the necessary linking flags
in the command it runs. When an un-modified Libtool is run, all the
redundant library linking flags are automatically removed by Libtool. But
when the modified Libtool is run, all those flags will be used and the
program will be able to compile.

This bug was reported by Raúl Infante-Sainz and fixed with this help.

This fixes bug #61192.