Radial profile: --instd and --stdhdu error when space used before value
Until now, when the '--instd' option was called with a space between it and
its value (for example '--instd /path/to/fits'), the script would crash
with an "Unknown option --instd" error. It would only work with an '=' sign
between the option name and value (for example
'--instd=/path/to/fits'). However, with the '=' sign, TAB autocompletion
doesn't usually work, so many people may avoid it and confront this
confusing error. A similar problem occurred for the '--stdhdu' option that
specifies the HDU/extension of the standard deviation image.

With this commit, in the option definition for these two options has been
edited so when the user does not use '=' to separate the option and its
value, the script will run without any error.