Book: adding short options for std image/HDU in radial profile script
Until this commit, when specifying the standard deviation image/HDU for the
radial profile script, the short options ('-i' and '-d'; added in the
previous commit) were missing in the book.

With this commit, the short option names have been added in the book. In
the process, it was discovered that in some cases, the short and long named
options (which sould start with '@item' and '@itemx', were both '@item';
causing too much empty space in formatted prints). So with this commit, the
long option name was changed to '@itemx'.

Also, the previous commit fixed bug #61128, but this fact wasn't added to
the NEWS file, so with this commit, we also added this bug as a fixed.