Table: wcs-to-img and img-to-wcs new names for wcstoimg and imgtowcs
Until now, the 'wcstoimg' and 'imgtowcs' were not following the standard
unit conversion naming convention as others (for example 'counts-to-mag' or
'jy-to-counts'). This can confuse users and is primarily for historical
reasons: the two WCS conversion operators were the first such
unit-converting operators, but after their usage, for later operators a
'-to-' was added to make the newer operators more easily readable.

With this commit, these two operators have been renamed to 'wcs-to-img' and
'img-to-wcs' in the code and documentation.