Library (units & arithmetic): added distance (LY-PC-AU) conversions
Until now, there was no easy way to do conversions between standard
distance measurements in astronomy. Users would have to have the conversion
factors in mind (or search them in the internet!) to write a generic
conversion in their script. But this was inconvenient!

With this commit, six functions have been added to the Units ('units.h')
library which are used by the Arithmetic library to define six new
operators for all types of conversion between the three commonly used
distance measures in astronomy: Light-years (LY), Parsecs (PC) and
Astronomical Units (AU).

While adding the items in the NEWS file, I noticed that I had forgot to
thank Ignacio Trujillo for the new feature of CosmicCalculator, so that was
also added.

This task was implemented after a question raised by Markus Schaney.