Book: fully working example on how to construct and correct the PSF
With this commit, I have added a fully working example that shows how to
construct and correct for the PSF.  It assumes an input image. Then, a Gaia
catalog is downloaded. The PSF related scripts are then used in order to
construct the PSF and model the scattered light field. The aim of this
example is to have a complete template to work with. Many details and
options are not used (like for example the segmentation images). But at
least we have a solid skeleton to start with.