PSF stamp: enabling the use of sigma-clip parameters
Until this commit, the option for using sigma-clip parameters was
available. However, it was not set properly inside of the script. In
practice that means that the option was not used. With this commit, this
bug has been fixed and now it is possible to use sigma-clip parameters when
using the script 'psf-create-make-stamp'.