PSF stamp: allowing to run the script without making normalization
Until this commit, when running the script 'psf-create-make-stamp', it
would abort as soon as the user doesn't provide a normalization radii with
the option '--normradii'. The reason is that this script is desired to
obtain normalized stamps of stars. However, it could be possible that the
user want the stamps without making the normalization.  For example,
because there is any other methodology for making the normalization. But at
the same time, it would be good to have this script running in order to
obtain the stamps (center the star, mask nearby objects, etc.). With this
commit, necessary changes have been included with the goal of running this
script without having to specify a ring of normalization. Internally, it
won't compute the radial profile and the normalization value will be set to
1.0.