PSF stamp: computing radial profile up to the minimum possible value
Until this commit, when creating the stamps of for constructing the PSF,
the radial profile was computed up to the maximum possible value. In
principle I thought it was not a problem since the stamps are allways
smaller than the original image. However, when trying to compute an
extended PSF, the size of the individual stamps could be considerable
large. And consequently, computing the entire radial profile is a waste of
CPU time (the radial profile is only necessary for making the
normalization). With this commit, this inefficiency has been fixed by
computing the radial profile up to a radius equal to the outer part of the
normalization ring. Since the radial profile is computed for the
normalization, it is not necessary to obtain larger radial profile than
this. However, if the user wants to obtain radial profile larger than this
default value, it is possible by using the --rmax option. Possible
situations are those in which the user want to have a look and check
individual radial profiles of all star stamps.