PSF stamp: add --saturation option for masking the saturated pixels
Until now, just the main stars had been cropped and extra objects around
them masked and finally normalized it. Without considering the saturated
pixels in the center of stars. In this situation the user do the
normalization mistakenly, because in the normalization, the statistic value
in each radius, are obtained mistakenly.
In the mask step, two 'astarithmetic' had been used.

Now, in the mask step instead of two 'astarithmetic', one 'astarithmetic'
has been used. Which technicaly the mask step will be done faster.
If the user wants to mask the saturation pixels, he/she can use
'--saturation' option and then mask the saturated pixels and if the user
does not determined this option the saturated pixels are not masked.