PSF model: computing radial profile up to the minimum possible value
Until this commit, when computing the flux factor, the radial profile was
computed up to the maximum radius that fits in the cropped image. However,
if the entire radial profile is not necessary, this is a waste of CPU time.

With this commit, this inefficiency has been fixed by computing the radial
profile up to a radius equal to the outer part of the normalization ring.
If the user wants to obtain radial profile larger than this default value,
it is possible by using the --rmax option.