PSF model: using PSF radial profile as input for flux factor computation
Until this commit, when computing the flux factor, only the PSF image was
provided. As a consequence, each time this script is executed, two radial
profiles have to be computed: the PSF and the object radial profiles.
However, in many situations, the PSF would be the same for many different
calls to this script. So, it would be good to have the possibility of
passing the radial profile of the PSf directly to the script to avoid
computing it. In this case, the radial profile has to be computed outside
of the script.

With this commit, this new feature has been added. Now the script is able
to use the radial profile of the PSF directly, without having to compute
it.