PSF stamp: allowingto use a position angle and axis ratio
Until this commit, when creating the stamps for the PSF, it was not
possible to use a given position angle nor axis ratio than the default
ones. So, the normalization was done on circulary symmetric apertures.
However, it is a good feature to allow changing the axis ratio and position
angle (for example, if the PSF is not circulary symmetric).

With this commit, this new feature have been added. Now the user is able to
specify a position angle and axis ratio for obtaining the radial profiles
and the normalization in a more flexible way.