PSF stamp: bug in --rmax option introduced in previous commit fixed
When rebasing the work of Sepideh over my branch, I changed the way of
computing the maximum radius for the normalization.  Sepideh did the
computation properly but I modified in a bad way her job. With this commit,
I have fixed this problem. Now the script doesn't use the '--rmax' option
anymore.