Library (arithmetic.h): sum operator uses pixels in data with blank
Until now, when the input to the 'sum' operator of the arithmetic library
(and thus the Arithmetic program as well as column arithmetic in Table) was
a floating point image that had blank values, all its pixels would be
ignored!

After looking into the code, I noticed that the problem was in the part
that we check for blank images: instead of comparing each element with
itself (to identify the NaN nature of the pixel), it was mistakenly always
comparing with the first pixel! So unless a pixel value was non-NaN and it
was exactly equal to the first image pixel, it would be ignored (assumed as
blank!) in the summation process.

With this commit, this coding typo has been corrected and the 'sum'
operator works properly.

This bug was found and reported by Giulia Golini and Raúl Infante-Sainz.

This fixes bug #61244.