Book: fixing references and labels for the PSF scripts information
Until this commit, the references and labels that are necessary for texinfo
were not properly included. With this commit, with the help of a couple of
Emacs macros, this issue has been solved.  In order to set the proper
references and labels I open 'doc/gnuastro.texi' with Emacs. Then I typed:

ctr-c, ctr-u, ctr-e      ---> To fix node items in texinfo.
ctr-u, ctr-c, ctr-u, m   ---> To update menues.