Book: extended introduction added to bitwise operators
Until now, there was no introductory explanation on the conditional,
bitwise, and building new dataset operators. While the first and third were
pretty easy to understand for a random astronomer, the second (bitwise
operators) could be a little hard to understand, so it was necessary to
give a good introduction in astronomical data analysis scenarios.

With this commit, an introduction has been added for all three sections and
the bitwise operators now have an extended discussion on why they are
useful for a modern astronomer.