PSF create: including the make check step for two PSF scripts
Until this commit, the 'make check' step of the scripts related to the
construction of the PSF were not set. With this commit, this step has been
added for two scripts (junction and stamps).  Now, during the installation
of Gnuastro they are check in order to ensure eveything is fine.