PSF model: including the make check step for two PSF scripts
Until this commit, the 'make check' step of the scripts related to the
modeling of objects using the PSF were not set. With this commit, this step
has been added for two scripts (flux-factor and scattered-light).  Now,
during the installation of Gnuastro they are check in order to ensure
eveything is fine.