Scripts: changing 'wcstoimg' to 'wcs-to-img' in several scripts
Until this commit, several scripts had the old notation for the operator
'wcstoimg'. However, this operator has been recently changed and now it is
'wcs-to-img'. See commit: 4e7edd16.  With this commit, this operator has
been changed in all the scripts so they follow now the new naming
convention. This issue was reported by Giulia Golini.