Development: built shell completion files added to .gitignore
Until now, these two files were not added to .gitignore because they were
created in a separate build directory when using
'developer-build'. However, when I did the './configure', 'make', 'make
check' and 'make install' within the source directory, I noticed these
files showing up as untracked files by git. But they are built files and
should not be included in Git at all.

With this commit, they have been added to '.gitignore' so now they are not
under Git control anymore and won't show up when Gnuastro is built in the
source directory.