Library (wcs.c): rename array to list converter
Until now, the 'wcs_convert_list_to_array' took a 'listtoarray'
argument. If one, it would indeed convert a list to an array, and vice
versa. However, since the function could convert both types to each other,
the '..._list_to_array' in the function name could lead to confusion.

With this commit, the function is renamed '..._list_to_from_array' to
indicate the capability to convert both ways.

Also, the parameter is renamed to 'to0from1' to facilitate the
understanding of which integer should the developer provide for which way
of conversion. For that matter, I had to swap the procedure inside if to
cope with this new naming convention.

Think of it this way: wcs_convert_list_to_from_array has 'to' as its first
use (the '0' as zero-indexed), and 'from' as its second function, hence the
'1' as its value.