Library (fits.h): better warning message in gal_fits_key_write_filename
Until now, the warning message just mentioned that the file name will be
truncated. But it didn't clarify that this is just metadata and won't
affect the analysis or output data. Also, it didn't say how the message can
be supressed.

With this commit, the warning message has been expanded to mention all
these points and make it more obvious to the user.

This was suggested by Sepideh Eskandarlou.