MakeCatalog: added breaks for two switch items that were missing it
Until now, two of the 'switch' blocks in MakeCatalog were missing a 'break'
when we were defining the column information to parse (related to the
'--geoz' and '--clumpsgeoz' columns). This could result in printing extra
columns for the user that they hadn't asked for (when they asked for any of
these columns).

With this commit, the 'break' statements have been added.

This fixes bug #61287.