PSF junction: fixed bug when centering the different parts of the PSF
Until this commit, when merging two different parts of the PSF the center
of the final merged PSF was not the correct one. In addition to this, also
the image used for masking the inner part of the outer PSF, and the outer
part of the inner PSF, was not properly constructed.  With this commit,
both problems have been solved. By changing the way the center of the PFS
and the mask profile, now the final PSF has the correct center position and
the junction is also done in the proper radius.