make check: using printf instead of SED in 1D convolution check
Until now, to feed the 1D kernel into Convolve's 1D convolution feature on
a test spectrum, we were using 'echo "1 3 10 3 1" | sed 's/ /\n/g''. But
after the build of Gnuastro on macOS of CondaForge (after the new release
0.16), I noticed that it crashed with this error:

    astconvolve: '1n3n10n3n1' couldn't be read as a number (element 1
    of first uncommented line)
    FAIL convolve/spectrum-1d.sh (exit status: 1)

With this commit, to fix the problem, instead of the pipe above, we are
simply using 'printf '1\n3\n10\n3\n1\n'' to avoid portability problems with
SED on different OSs.

This fix was implemented with the help of Sebastian Luna-Valero (while
testing on CondaForge's automatic build [1]).

This fixes bug #61329.

[1] https://github.com/conda-forge/gnuastro-feedstock/pull/14