Book: Correct the some parts of the making the PSF
Until now, in the first part of the book about the making PSF, the stars
have been selected by 'astquery' from Gaia and then by using 'asttable'
sort all of stars and just select just 5 brightness of them. And the
comment in the book was so long.

Now, for selecting the stars in different range of magnitude for making the
different part of the PSF, we do not use the 'astquery' and 'asttable' for
selecting the bright stars. I explained that just by using
'astscript-create-select-stars' we can select stars in the special range of
magnitude and long commite are shrinked.

Now, some parts have corrected.