PSF select-stars: changing some variable names to follow conventions
Until this commit, this script had several variable names that come from
long time ago and they did not follow the conventions we are using in
Gnuastro (at least, for the scripts developed for the PSF). With this
commit, I have corrected some of them. Still, the script needs to be
revised to ensure that it works properly.