Segment: --noobjects is new name for old --onlyclumps
Until now, the '--onlyclumps' option name could cause a confusion: leading
the user to assume that the output will _only_ contain the 'CLUMPS' HDU!
But this was not the case by default! To get such an output, the user would
also have to give '--rawoutput'.

With this commit, to avoid such confusions, the option name has been
changed to '--noobjects'.

This issue was reported by Sepideh Eskandarlou.