PSF select-stars: removing not used options and polishing the comments
Until this commit, there were some options that the script was not using.
For example the '--show' option. With this commit, this option and other
parts of the script that were not used have been removed. In addition to
this, several parts like the comments and descriptions have been changed to
be more clear and precise.