PSF scripts: checking that the input file actually exists
Until this commit, all PSF related scripts checked if the 'inputs' variable
was empty or not. If it was empty, an error and message was printed to
warn the user that no input was provided. However, if the input file was
not properly specified (i.e., a typo while writing it), the scripts tried
to continue without checking if the inputs actually exists. This is bad,
and the script should stop as soon as it detects that the input file does
not exists. With this commit, this bug has been fixed for several scripts.

This problem was reported by Sepideh Eskandarlou.