Book: polishing the description of the PSF select-stars script
Until this commit, the generation of the book in PDF format was not
possible because some references were missing. With this commit, I have
fixed these references with the help of emacs routines as described in
commit d3a69fad. In addition to this, several parts of the Book has been
changed in order to be more clear in the description of the
psf-create-select-star script.