Book: new section for the tutorial on how to construct and extended PSF
Until this commit, the working example on how to construct the extended PSF
using the script we have develop was too long (in a single block of code).
This is not good for the new users because it is hard to read.

With this commit, I have slpited it into different blocks of codes with a
description. In addition to this, the entire tutorial has been put into a
dedicated section just before the description of the different PSF scripts.
In the same way, the comment above some examples on how to run the scripts
have been reduced to only have two or three lines.