PSF select-stars: corrected bug in checking the catalog-image overlap
Until this commit, there was an error when checking the overlapping between
the objects of the catalog and the border of the input image. The problem
was basically that the condition was checking if the objects were outside
of the image, but it has to be in completely the contrary: we want the
catalog of objects to be within the image. So, the minimum RA/DEC of the
catalog has to be _larger_ than the minimum RA/DEC of the image. In the
same way, the maximum RA/DEC of the catalog has to be _smaller_ than the
minimum RA/DEC of the image. I discovered this problem when trying to
include the make checks for this script.

With this commit, this bug has been fixed by changing the conditions in
this check.