PSF select-stars: changing default values to not crash in make check
Until this commit, there were some parameters set to by default values that
don't allow to run the make check without a crash. With this commit, I have
changed them to be able to run the make check without any problem. Of
course, this has to be corrected. Otherwise the script result will have no
sense in some situations. The necessary fix is to not assume any value by
default, and force the user to specify them. For example, axis ratio,
magnitude, colum names for the coordinates, etc.