PSF select-stars: The script has been changed for use in general scenario
Until now, for '--min' and '--max' range of magnitude, default values were
considered. The table columns were not general and they had fixed name which
corresponded to the Gaia tables such as ra, dec, parallax and
parallax_error. The output name of the some files was incorrect.

With this commit, the '--min' and '--max' options have removes and instead
of them the '--magnituderange' option is added and based on that the user
can give two values (minimum and maximum range of magnitude) to this
option. For using this script in the general scenario some options such as
'--racolumn', '--deccolumn' and '--parallaxanderrorcolumn' option are added
to this scrip, which based on that the user can determine the column
name. That may differ from the names of the columns in Gaia. Some outputs
name has changed.