Book: some essential option was add to astscript-psf-select-stars section
Until now, for selecting the stars in the range of magnitude use two
different option and there was no option for the name of the columns.

With this commit, the '--min' and '--max' options are removed and instead of
them the '--magnituderange' has been added and explained about it. The
'--racolumn', '--deccolumn', '--parallaxanderrorcolumn' are for names of the
columns.