PSF select-stars: delete the value of some options and ask them from the user
Until now, some options such as '--minaxisratio', '--mindistdeg', and
'matchaperturedeg' had a value by default and if the user did not provide
these values the script had been working. While these options are essential
values in 'astscript-create-select-stars' script and if the user does not
provide these options the script worked with the by default values and get
the mistake results. And in some scenario may be the user do not want to
select stars with good parallax, while in astqurey she/he should provide
this column.

With this commit, the value of these options removed. And if the user does
not provide these options the script will be intterupt and print an
error. And if the user do not want provide the stars with good parallax she/he
can easily did not determined this '--parallaxanderrorcolumn' option.