Library (txt.c): no error when sexagesimal used for column info
Until now, when a plain-text table with no metadata contained sexagesimal
values in the first row (that is used to extract metadata for the rest of
the rows), Table would crash with an error saying that the sexagesimal
number couldn't be interpretted as a number! As a result a basic command
like below (which is even in the "Invoking asttable" section of the book),
would fail:

    echo "7h34m35.5498 31d53m14.352s" | asttable

This had happened because of a recent change in the plain-text reading
library that was added to make sure that the first un-commented line of a
file without metadata actually has numbers in it (bug #60999).

With this commit, that part of the code has been corrected to also check
for sexagesimal inputs, not just raw numbers. Also, to make sure that this
situation is tested every time, the command above has been added to the
'make check' tests.

This fixes bug #61378.