PSF select-stars: check the image and catalog have a overlap area or not
Until now, if the user provides a catalog of stars and want based on that
and input image, select the stars in the specific range of the magnitude,
just check the image are smaller than the catalog or larger than it. But
did not check that image and catalog have an overlap area together or not.

With this commit, before selecting the stars in the specific range of
magnitude, at first by using the 'astfits' and '--skycoverage' option we
find the four coordinate of the iamge, then based on these coordinate and
by using the '--polygon' option in 'asttable' the overlap area between the
image and the catalog has been found and if they have not to overlap area
(if the number of rows in table to be zero) the script will be stopped and
print an error, otherwise, if the number of rows in this table not to be
zero the script will be run and select the stars in the range of magnitude
between minimum and maximum.