Book: added section on measuring the magnitude limit of an image
Until now, the "Quantifying magnitude limits" section of the book didn't
include the magnitude limit of the dataset at a certain S/N. This is also
used by some studies to quantify "depth", and can be mentioned as
"magnitude limit" or "detection limit".

With this commit, a section describing this measure has been added to the
book and the caveat of it (how it depends on the morphology of the objects)
is also discussed.