Book: all output extensions of --checkqthresh now described
Until now, this very important option in NoiseChisel didn't have any proper
description of what each output extension is. This would discourage people
from using it (because of the many extensions, it could be hard to
understand).

With this commit, each extension of the output file with '--checkqthresh'
is described and the options that affect that extension are described.

This issue was reported by Zahra Hosseini.