PSF select-stars: make check steps done again because some new option are added
Until now, the '--magnituderange', '--mindistdeg', '--minaxisratio',
'--matchaperturedeg' options did not exist and 'make check' steps was set
for the selecting of the stars without these options. Without considering
these options in 'make check' steps the fail ocuured.

With this commit, the '--magnituderange', '--mindistdeg', '--minaxisratio',
'--matchaperturedeg' have been considered in the 'make check' steps and now
in the 'make check' steps the user do not faced with a fail. And during the
installation of the Gnuastro this script will be checked to make sure
everything is fine.