Book: add magnitude quantile to the surface brightness limit section
Until now, the definition of skewness was used to quantitatively compare
signal and noise distribution in surface brighness limit section.

Now, for quantifying the signal and noise distributions instead of skewness
we used from quantile of the mean. Because the distribution of the signal
and the noise are not Gaussian and we cant used the skewness, we should use
the quantile of the mean instead of skewness.