Match: k-d tree matching UI fixed, to do: finalizing docs and tests
Until now, while most low-level aspects of the k-d tree based matching were
done, the high-level aspects and in paritcular the documentation wasn't
done yet.

With this commit, the UI now works in the various scenarios and the
documentation is mostly complete (in describing the algorithms). However,
the options are still not added and no tests have been added to 'make
check' yet.