Library (match.h): k-d tree method discards regions with no overlap
Until now, the k-d method would attempt to parse the full k-d of the first
catalog for each item of the second catalog. When the two catalogs
spatially overlap significantly this is fine (because the k-d tree will
find the nearest neighbor pretty fast and we can reject it if it is out of
the user's given aperture).

However, when the two datasets don't significantly overlap (and catalog B
covers a much larger spatial area than catalog A), this can be very
inefficient. Because the k-d tree parsing algorithm will be forced to parse
the whole tree for every point! I encountered such a case while testing the
Mach program on some large tables (390165 and 5773552 rows respectively)
provided by Andres Del Pino Molina.

With this commit, a solution to this problem has been implemented: before
attempting the parallelization of the B catalog rows, we will first obtain
A's histograms in each dimention. Later, when parsing the rows of catalog
B, we will first check if the point in B actually falls in a non-zero bin
of the N-dimensional histogram. If it doesn't we won't even bother calling
the k-d tree algorithm. The discussion on the matching algorithms in the
book now also mentions this new aspect of the k-d tree matching.

This resulted in a MAJOR speed improvement for the catalogs given by
Andres: from about +20 minutes (we stopped it manually!), to less than 20
seconds!