Match: matched rows aren't permuted, but new column allocated
Until now, once the matched rows were found, the full input columns were
permuted. A a result, when the number of rows was large, it would be very
slow (many of the un-necessary rows were permuted with the good ones).

With this commit, instead of permutation, we simply allocate a new array
and copy the matched items into it. This resulted in a great speed
increase. Also, since the row copying is independent between rows, we are
now doing the copying in parallel.