Library (fits.h): reading table columns now done in parallel
Until now, the column reading of FITS tables was done in series (one after
the other). This was because CFITSIO reads each column independently and we
were using a simple 'for' loop to parse it. As a result, when the table had
many rows and many columns, simply reading the columns/rows could become
very slow!

With this commit, the function to open FITS files and read the table
columns has been parallelized, allowing all requested threads to read from
the file at one instance. This resulted in a great speed increase when
reading large tables.

As a result of this, it is was necessary to add a new argument 'numthreads'
to the two 'gal_table_read' and 'gal_fits_tab_read' functions of the
library.