Library (fits.h): corrected segmentation fault in parallel reading
Until now (in the code that was added in the previous commit), I had
mistakenly used 'i' in two loops within each other (one for parsing the
columns given to that thread, another for parsing string columns). So when
the input had string columns, the library would crash with a segmentation
fault.

With this commit, a separate 'j' variable was used for the the inner loop
and this fixed the problem.