Library (fits.h): correctly reading columns of 0 width/repeat
Until now, if a column of a FITS binary table had a 0 width or "repeat"
value (in the TFORM keyword, for example '0A' in for a 0-repeat string
column), Gnuastro would crash with the following CFITSIO error:

   FITSIO status = 308: bad first element number
   First element to write is too large: 1; max allowed value is 0

This was because until now, we were assuming that a column always has
atleast one element!

With this commit, upon reading the table's metadata (and discovering that
it has a "repeat" value of 0), a bit flag is set so when we actually want
to read the column, in such cases, it will simply read an all-blank set of
values (which would be the equivalent value in Gnuastro's internal data
structure, for now).

In the process the following changes are also made:

 - Until now, there was only a single column metadata flag, so we had used
   an 'enum' and '==' to check for it. But with this commit, we need more
   than one flag, so the flags are now bit flags and all relevant checks
   have been corrected.

 - This whole process happened while I was editing the Matching part of the
   Gnuastro book, so those edits are also being committed (they are just
   minor copy-edits, and not worth a separate commit!).